requirements.txt
# Django - related
django
django-debug-toolbar
djangorestframework
django-cors-headers
markdown       # Markdown support for the browsable API of djangorestframework.
django-filter  # Filtering support - for DRF



#  DB drivers
mysqlclient

# reads .env, and converts to env variables
python-dotenv

# Deploy
gunicorn
getenv
